Add Summary Table
Table contains information for each database added to the summary page
accessible through the existing summary statistics in the header, the total number of open
issues, total number of found issues, total number of fixed issues.

The backend function to facilitate this table was added to the server
route.

Removed previous template CRUD functions in server.