Update UI for issues page
Since MRT already has its own elevation, adding another layer of
elevation beneath it is redundant. The refresh button on the header is
also unnecessary, as there is already a refresh button on the top left
of the table.

Let's remove the styled card around the table and the refresh button on
the header to keep things simple.