Load Bug Reports From Sqlancer Github Bugs Page
The backend will create a new database table and load in the data into
that new table if it is empty on server start.

The schema of this new table includes some other new fields but the
columns dbms and status remains consistent for the sake of summary
aggregation.

The 2 methods which fetches summary data has been updated to include the
data from this new table through the creation date, status and dbms.

A new frontend page has been created with the same table that shows the
other fields which exists on this table. This table only supports edits
for the status column since all reports here has already been vetted by
the prof. There has been no updated to this bugs.json page for over a
year hence there are no functions to refetch any data.